requirements.txt
Flask==2.3.3
mysql.connector==8.1.0
json==2.0.9